Bug 1963833: don't attempt to delete nil PodDisruptionBudget object
NewPodDisruptionBudget may return nil PodDisruptionBudget when !f.infrastructure.HighlyAvailableInfrastructure(),
this would cause a segfault later when DeletePodDisruptionBudget was
called with that nil PodDisruptionBudget, causing Single Node clusters
to fail to install.

+++ pkg/tasks/prometheus_user_workload.go
@@ -308,9 +308,11 @@ func (t *PrometheusUserWorkloadTask) destroy() error {
 		return errors.Wrap(err, "initializing UserWorkload Prometheus PodDisruptionBudget object failed")
 	}
 
-	err = t.client.DeletePodDisruptionBudget(pdb)
-	if err != nil {
-		return errors.Wrap(err, "deleting UserWorkload Prometheus PodDisruptionBudget object failed")
+	if pdb != nil {
+		err = t.client.DeletePodDisruptionBudget(pdb)
+		if err != nil {
+			return errors.Wrap(err, "deleting UserWorkload Prometheus PodDisruptionBudget object failed")
+		}
 	}
 
 	err = t.client.DeleteSecret(s)
